Changes to update from Rust 1.7 to 1.10
Most of the changes are for unstable features I was using (the panic
stuff is now stable). I also had to add a cast before transmuting
function pointers. The compile warning linked to
https://github.com/rust-lang/rust/issues/19925

+++ src/dialog.rs
@@ -150,7 +150,7 @@ impl ShowState {
 callback_token!(ShowCallbackToken);
 thread_local!(
     static SHOW_CALLBACKS: CallbackRegistry<FnMut(ShowState) -> CallbackAction, ShowCallbackToken> =
-        CallbackRegistry::new("SHOW_CB", unsafe { mem::transmute::<_, Icallback>(show_cb) })
+        CallbackRegistry::new("SHOW_CB", unsafe { mem::transmute::<_, Icallback>(show_cb as usize) })
 );
 extern fn show_cb(ih: *mut Ihandle, state: c_int) -> c_int {
     with_callbacks(ih, &SHOW_CALLBACKS, |cbs| {
